refactor: organize shader creation part
